Added Tests. Services coverage =100%, xmlWriter =88%
