+++ uiServices/services_test.go
@@ -2,32 +2,15 @@ package services
 
 import (
 	"testing"
-	"github.com/stretchr/testify/assert"
 	"os"
 	"net/http/httptest"
 	"strings"
 	"github.com/go-chi/chi"
 	"io"
 	"net/http"
+	"github.com/Sirupsen/logrus"
+	"github.com/xtracdev/automated-perf-test/perfTestUtils"
 )
-
-
-
-func TestFilePathExist(t *testing.T) {
-	path := os.Getenv("GOPATH")
-	actual := false
-	actual = FilePathExist(path)
-	expected := true
-	os.Getenv("GOPATH")
-	assert.Equal (t, expected, actual)
-}
-
-func TestFilePathDoesNotExist(t *testing.T) {
-	path := "C:/Users/a615194/go/src/github.com/xtrac"
-	actual := FilePathExist(path)
-	expected := false
-	assert.Equal (t, expected, actual)
-}
 
 
 var (
@@ -36,13 +19,11 @@ var (
 
 
 func TestValidJsonPost(t *testing.T) {
-	//start server
 	r := chi.NewRouter()
 	r.Mount("/", getIndexPage())
-
-	//create JSON
+	
 	configJson := `{
-        "apiName": "TestConfigFile",
+        "apiName": "TestConfig",
        "targetHost": "localhost",
        "targetPort": "9191",
        "numIterations": 1000,
@@ -60,12 +41,14 @@ func TestValidJsonPost(t *testing.T) {
        "rampUsers": 5,
        "rampDelay": 15
        }`
-	//reader
+
 	reader= strings.NewReader(configJson)
-	//prepare POST method
 	r.HandleFunc("/configs", configsHandler)
-	//perfrom POTS method using URI and read JSON
+
+	filePath :=os.Getenv("GOPATH")+"/src/github.com/xtracdev/automated-perf-test/config/"
+
 	request, err := http.NewRequest(http.MethodPost, "/configs", reader)
+	request.Header.Set("configPathDir", filePath)
 
 	w := httptest.NewRecorder()
 	r.ServeHTTP(w, request)
@@ -76,15 +59,15 @@ func TestValidJsonPost(t *testing.T) {
 
 	if http.StatusCreated != 201 {
 		t.Errorf("Error. Did not succesfully post")
+	}else{
+	logrus.Print("Successfully created XML File")
 	}
 }
 
 func TestInvalidJsonPost(t *testing.T) {
-	//start server
 	r := chi.NewRouter()
 	r.Mount("/", getIndexPage())
-
-	//create JSON
+	
 	configJson := `{
         "apiName": 0,
        "targetHost": 0,
@@ -104,11 +87,10 @@ func TestInvalidJsonPost(t *testing.T) {
        "rampUsers": 5,
        "rampDelay": 15
        }`
-	//reader
+	
 	reader= strings.NewReader(configJson)
-	//prepare POST method
 	r.HandleFunc("/configs", configsHandler)
-	//perfrom POTS method using URI and read JSON
+
 	request, err := http.NewRequest(http.MethodPost, "/configs", reader)
 
 	w := httptest.NewRecorder()
@@ -123,12 +105,89 @@ func TestInvalidJsonPost(t *testing.T) {
 	}
 }
 
+func TestPostWithNoFilePath(t *testing.T) {
+	r := chi.NewRouter()
+	r.Mount("/", getIndexPage())
 
-//
-//func TestFileSave(t *testing.T) {
-//
-//	path :=  "C:/Users/a615194/go/src/github.com/xtracdev/automated-perf-test/config/ConfigTest.xml"
-//	actual := FilePathExist(path)
-//	expected := false
-//	assert.Equal (t, expected, actual)
-//}
+	configJson := `{
+        "apiName": "Test",
+       "targetHost": "localhost",
+       "targetPort": "9191",
+       "numIterations": 1000,
+       "allowablePeakMemoryVariance": 30,
+       "allowableServiceResponseTimeVariance": 30,
+       "testCaseDir": "./definitions/testCases",
+       "testSuiteDir": "./definitions/testSuites",
+        "baseStatsOutputDir": "./envStats",
+       "reportOutputDir": "./report",
+       "concurrentUsers": 50,
+       "testSuite": "suiteFileName.xml",
+       "memoryEndpoint": "/alt/debug/vars",
+       "requestDelay": 5000,
+       "TPSFreq": 30,
+       "rampUsers": 5,
+       "rampDelay": 15
+       }`
+
+	reader= strings.NewReader(configJson)
+
+	r.HandleFunc("/configs", configsHandler)
+	request, err := http.NewRequest(http.MethodPost, "/configs", reader)
+
+	w := httptest.NewRecorder()
+	r.ServeHTTP(w, request)
+
+	if err != nil {
+		t.Error(err)
+	}
+}
+
+func TestInvalidURL(t *testing.T) {
+	pt := perfTestUtils.Config{}
+	writerXml(pt, "/path/xxx")
+}
+
+func TestInvalidFileName(t *testing.T) {
+	r := chi.NewRouter()
+	r.Mount("/", getIndexPage())
+
+	configJson := `{
+        "apiName": "//*()()()",
+       "targetHost": "localhost",
+       "targetPort": "9191",
+       "numIterations": 1000,
+       "allowablePeakMemoryVariance": 30,
+       "allowableServiceResponseTimeVariance": 30,
+       "testCaseDir": "./definitions/testCases",
+       "testSuiteDir": "./definitions/testSuites",
+        "baseStatsOutputDir": "./envStats",
+       "reportOutputDir": "./report",
+       "concurrentUsers": 50,
+       "testSuite": "suiteFileName.xml",
+       "memoryEndpoint": "/alt/debug/vars",
+       "requestDelay": 5000,
+       "TPSFreq": 30,
+       "rampUsers": 5,
+       "rampDelay": 15
+       }`
+
+	reader= strings.NewReader(configJson)
+
+	r.HandleFunc("/configs", configsHandler)
+	request, err := http.NewRequest(http.MethodPost, "/configs", reader)
+
+	filePath :=os.Getenv("GOPATH")+"/src/github.com/xtracdev/automated-perf-test/config/"
+	request.Header.Set("configPathDir", filePath)
+
+	w := httptest.NewRecorder()
+	r.ServeHTTP(w, request)
+
+	if err != nil {
+		t.Error(err)
+	}
+
+	if http.StatusCreated != 201 {
+		t.Errorf("Error. Did not succesfully post")
+	}
+
+}
